requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple

